--- packages/storybook/examples/core/Button/ExpandedButton.js
@@ -1,28 +1,0 @@
-import React from 'react';
-
-import Button from '@ichef/gypcrete/src/Button';
-import FlexRow from 'utils/FlexRow';
-
-function ExpandedButtonExample() {
-    return (
-        <div>
-            <FlexRow>
-                <Button
-                    basic="Expanded Button"
-                    align="center"
-                    minified={false} />
-            </FlexRow>
-
-            <FlexRow>
-                <Button
-                    solid
-                    color="red"
-                    basic="Expanded Button"
-                    align="center"
-                    minified={false} />
-            </FlexRow>
-        </div>
-    );
-}
-
-export default ExpandedButtonExample;
